requirements.txt
jupyter-book
matplotlib
numpy
pandas
scipy.stats
seaborn